+++ src/main/java/com/woozuda/backend/note/service/QuestionNoteService.java
@@ -7,16 +7,15 @@ import com.woozuda.backend.diary.repository.DiaryRepository;
 import com.woozuda.backend.note.dto.request.QuestionNoteModifyRequestDto;
 import com.woozuda.backend.note.dto.request.QuestionNoteSaveRequestDto;
 import com.woozuda.backend.note.dto.response.NoteResponseDto;
-import com.woozuda.backend.note.entity.CommonNote;
 import com.woozuda.backend.note.entity.NoteContent;
-import com.woozuda.backend.note.entity.Question;
+import com.woozuda.backend.question.entity.Question;
 import com.woozuda.backend.note.entity.QuestionNote;
 import com.woozuda.backend.note.entity.type.Feeling;
 import com.woozuda.backend.note.entity.type.Season;
 import com.woozuda.backend.note.entity.type.Weather;
 import com.woozuda.backend.note.repository.NoteRepository;
 import com.woozuda.backend.note.repository.QuestionNoteRepository;
-import com.woozuda.backend.note.repository.QuestionRepository;
+import com.woozuda.backend.question.repository.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
